Community 페이지-search form 분리, search input 초기화 구현
