+++ src/components/community/CommunityArticleList.jsx
@@ -1,0 +1,7 @@
+function CommunityArticleList({ data }) {
+    return (
+        <div>리스트</div>
+    );
+}
+
+export default CommunityArticleList;
