+++ src/components/community/index.js
@@ -1,3 +1,5 @@
 export { default as Category } from './Category';
+export { default as SearchForm } from './SearchForm';
 export { default as OrderList } from './OrderList';
 export { default as RegionDropdown } from './RegionDropdown';
+export { default as CommunityArticleList } from './CommunityArticleList';
