+++ src/styles/components/_input.scss
@@ -40,6 +40,7 @@
         }
 
         &--no-line {
+            margin-left: 0.5rem;
             border: none;
         }
     }
