+++ src/styles/pages/_community.scss
@@ -35,6 +35,7 @@
         display: flex;
         justify-content: space-between;
         align-items: center;
+        margin-bottom: 1rem;
     }
 
     &__write-button {
